requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
frozenlist==1.7.0
greenlet==3.2.3
idna==3.10
multidict==6.6.3
playwright==1.53.0
propcache==0.3.2
pyaes==1.6.1
pyasn1==0.6.1
pyee==13.0.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rsa==4.9.1
soupsieve==2.7
Telethon==1.40.0
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
yt-dlp==2025.7.21
